8.15. Открывается и закрывается (часть 2)
