+++ js/main.js
@@ -1,4 +1,2 @@
 import './photo/add-user-photo-main-page.js';
 import './photo/open-big-photo.js';
-import './photo/close-big-photo.js';
-
