+++ js/photo/add-user-photo-main-page.js
@@ -1,6 +1,5 @@
 import {descriptionPhoto} from './create-description-photo';
 
-
 const picturesList = document.querySelector('.pictures');
 const templatePicture = document.querySelector('#picture').content.querySelector('.picture');
