+++ Обучение 1.code-workspace
@@ -1,0 +1,13 @@
+{
+	"folders": [
+		{
+			"path": "../../Обучение 1"
+		},
+		{
+			"path": "."
+		}
+	],
+	"settings": {
+		"liveServer.settings.multiRootWorkspaceName": "Обучение 1"
+	}
+}
